+++ Baekjoon/silver/스택_10828/wooseok.java
@@ -1,14 +1,6 @@
-package Baekjoon.silver.m6w3;
+package Baekjoon.silver.스택_10828;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.Deque;
-import java.util.LinkedList;
-import java.util.Stack;
-import java.util.StringTokenizer;
-
-public class stack_10828_wooseok {
+public class wooseok {
 
     // 1. stack을 직접 사용
     /*
